# internal/server/gen/swagger.yaml
components:
  schemas:
...
    ent.ActionItem:
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        description:
          description: Description holds the value of the "description" field.
          type: string
        due_date:
          description: DueDate holds the value of the "due_date" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.ActionItemEdges'
        id:
          description: ID of the ent.
          type: integer
        owner_id:
          description: OwnerID holds the value of the "owner_id" field.
          type: string
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_actionitem.Status'
        tracker_url:
          description: 对接外部工单
          type: string
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
    ent.ActionItemEdges:
      description: |-
        Edges holds the relations/edges for other nodes in the graph.
        The values are being populated by the ActionItemQuery when eager-loading is set.
      properties:
        postmortem:
          $ref: '#/components/schemas/ent.Postmortem'
      type: object
...
    ent.Event:
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        dedup_key:
          description: 去重键
          type: string
        detail:
          additionalProperties: {}
          description: 原始 payload 明细
          type: object
        edges:
          $ref: '#/components/schemas/ent.EventEdges'
        id:
          description: ID of the ent.
          type: integer
        is_noise:
          description: 分诊判定为噪音
          type: boolean
        labels:
          additionalProperties:
            type: string
          description: 路由用标签
          type: object
        received_at:
          description: ReceivedAt holds the value of the "received_at" field.
          type: string
        severity:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_event.Severity'
        source:
          description: 告警源，如 prometheus
          type: string
        source_event_id:
          description: 原始告警 ID，去重 + 幂等键
          type: string
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_event.Status'
        summary:
          description: 一句话摘要
          type: string
      type: object
...
    ent.Incident:
      description: Incident holds the value of the incident edge.
      properties:
        acked_at:
          description: 确认时间，MTTA 计算
          type: string
        closed_at:
          description: ClosedAt holds the value of the "closed_at" field.
          type: string
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        current_level:
          description: 当前升级层级
          type: integer
        edges:
          $ref: '#/components/schemas/ent.IncidentEdges'
        embedding:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_schema.NullableVector'
        escalated_count:
          description: 已升级次数
          type: integer
        id:
          description: ID of the ent.
          type: integer
        merged_into:
          description: 若被合并，指向主 Incident
          type: string
        number:
          description: 人类可读编号，如 INC-0042
          type: string
        priority:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.Priority'
        resolved_at:
          description: ResolvedAt holds the value of the "resolved_at" field.
          type: string
        severity:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.Severity'
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.Status'
        summary:
          description: 当前概要，可随处置更新
          type: string
        title:
          description: Title holds the value of the "title" field.
          type: string
        trigger_source_event_id:
          description: TriggerSourceEventID holds the value of the "trigger_source_event_id"
            field.
          type: string
        trigger_type:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.TriggerType'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
        war_room:
          additionalProperties: {}
          description: 作战室：im_platform/im_channel_id/created_at
          type: object
      type: object
...
    ent.Integration:
      properties:
        config:
          additionalProperties: {}
          description: 类型相关配置
          type: object
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.IntegrationEdges'
        enabled:
          description: Enabled holds the value of the "enabled" field.
          type: boolean
        id:
          description: ID of the ent.
          type: integer
        name:
          description: Name holds the value of the "name" field.
          type: string
        type:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_integration.Type'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
    ent.IntegrationEdges:
      description: |-
        Edges holds the relations/edges for other nodes in the graph.
        The values are being populated by the IntegrationQuery when eager-loading is set.
      properties:
        events:
          description: Events holds the value of the events edge.
          items:
            $ref: '#/components/schemas/ent.Event'
          type: array
          uniqueItems: false
        raw_events:
          description: RawEvents holds the value of the raw_events edge.
          items:
            $ref: '#/components/schemas/ent.RawEvent'
          type: array
          uniqueItems: false
        service:
          $ref: '#/components/schemas/ent.Service'
        team:
          $ref: '#/components/schemas/ent.Team'
      type: object
...
    ent.Postmortem:
      description: Postmortem holds the value of the postmortem edge.
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.PostmortemEdges'
        embedding:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_schema.NullableVector'
        generated_by:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_postmortem.GeneratedBy'
        id:
          description: ID of the ent.
          type: integer
        published_at:
          description: PublishedAt holds the value of the "published_at" field.
          type: string
        sections:
          additionalProperties: {}
          description: 结构化内容章节
          type: object
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_postmortem.Status'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
    ent.PostmortemEdges:
      description: |-
        Edges holds the relations/edges for other nodes in the graph.
        The values are being populated by the PostmortemQuery when eager-loading is set.
      properties:
        action_items:
          description: ActionItems holds the value of the action_items edge.
          items:
            $ref: '#/components/schemas/ent.ActionItem'
          type: array
          uniqueItems: false
        incident:
          $ref: '#/components/schemas/ent.Incident'
      type: object
...
    github_com_kevin_vigil_ent_actionitem.Status:
      description: Status holds the value of the "status" field.
      enum:
      - open
      - in_progress
      - done
      type: string
      x-enum-varnames:
      - DefaultStatus
      - StatusOpen
      - StatusInProgress
      - StatusDone
...
    github_com_kevin_vigil_ent_integration.Type:
      description: Type holds the value of the "type" field.
      enum:
      - webhook
      - email
      - prometheus
      - zabbix
      - grafana
      - cloud
      - api
      type: string
      x-enum-varnames:
      - TypeWebhook
      - TypeEmail
      - TypePrometheus
      - TypeZabbix
      - TypeGrafana
      - TypeCloud
      - TypeAPI
...
    github_com_kevin_vigil_ent_postmortem.GeneratedBy:
      description: GeneratedBy holds the value of the "generated_by" field.
      enum:
      - mixed
      - ai
      - human
      type: string
      x-enum-varnames:
      - DefaultGeneratedBy
      - GeneratedByAi
      - GeneratedByHuman
      - GeneratedByMixed
    github_com_kevin_vigil_ent_postmortem.Status:
      description: Status holds the value of the "status" field.
      enum:
      - draft
      - in_review
      - published
      - archived
      type: string
      x-enum-varnames:
      - DefaultStatus
      - StatusDraft
      - StatusInReview
      - StatusPublished
      - StatusArchived
...
    github_com_kevin_vigil_ent_schema.NullableVector:
      description: 语义向量，published 复盘入库后计算，知识沉淀检索用
      properties:
        valid:
          type: boolean
      type: object
...
    github_com_kevin_vigil_internal_httputil.ErrorResponse:
      properties:
        code:
          description: Code 机器可读错误码（可选，非 4xx 通用错误可留空）。
          example: invalid_argument
          type: string
        details:
          description: Details 结构化补充信息（可选，如字段级校验错误）。
        error:
          example: invalid id
          type: string
      type: object
...
    internal_integration.createReq:
      properties:
        config:
          additionalProperties: {}
          description: 类型相关配置（URL/过滤/限流等）
          type: object
        name:
          type: string
        service_id:
          type: integer
        team_id:
          type: integer
        type:
          description: webhook|email|prometheus|zabbix|grafana|cloud|api
          type: string
      type: object
    internal_integration.createResp:
      properties:
        config:
          additionalProperties: {}
          description: 类型相关配置
          type: object
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.IntegrationEdges'
        enabled:
          description: Enabled holds the value of the "enabled" field.
          type: boolean
        id:
          description: ID of the ent.
          type: integer
        name:
          description: Name holds the value of the "name" field.
          type: string
        token:
          description: ★ 明文 token，仅创建时返回一次
          type: string
        type:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_integration.Type'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
    internal_integration.updateReq:
      properties:
        enabled:
          type: boolean
        name:
          type: string
      type: object
...
    internal_postmortem.addActionItemReq:
      properties:
        description:
          type: string
        owner_id:
          type: string
        tracker_url:
          type: string
      type: object
    internal_postmortem.transitionReq:
      properties:
        status:
          description: in_review | published | archived
          type: string
      type: object
    internal_postmortem.updateActionItemReq:
      properties:
        owner_id:
          type: string
        status:
          description: open | in_progress | done
          type: string
        tracker_url:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 业务 API 推荐（唯一声明）方案：HTTP Bearer JWT（/auth/login 换取）。
      scheme: bearer
      type: http
...
info:
  description: |-
    Vigil 告警处置平台 REST API。
    认证：业务 API 要求 Bearer JWT（/auth/login 换取）。鉴权中间件另接受 X-Vigil-User-ID 头作为本地/回退身份（可伪造，仅限受信网络，生产禁用），该回退方案不在 securitySchemes 中声明。
  license:
    name: MIT
  title: Vigil API
  version: 0.1.0
openapi: 3.1.0
paths:
  /action-items/{id}:
    patch:
      parameters:
      - description: 改进项 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_postmortem.updateActionItemReq'
                description: 更新字段（全部可选）
                summary: body
        description: 更新字段（全部可选）
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.ActionItem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 更新改进项
      tags:
      - postmortem
...
  /incidents/{id}/postmortem/draft:
    post:
      description: 基于时间线与（可选）AI 生成事件复盘草稿。
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Postmortem'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 生成复盘草稿
      tags:
      - postmortem
...
  /integrations:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ent.Integration'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 接入点列表
      tags:
      - integration
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_integration.createReq'
                description: 接入点配置
                summary: body
        description: 接入点配置
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_integration.createResp'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 创建接入点（返回 token 仅一次）
      tags:
      - integration
  /integrations/{id}:
    delete:
      parameters:
      - description: 接入点 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
      security:
      - bearerAuth: []
      summary: 删除接入点
      tags:
      - integration
    get:
      parameters:
      - description: 接入点 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Integration'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Not Found
      security:
      - bearerAuth: []
      summary: 接入点详情
      tags:
      - integration
    patch:
      parameters:
      - description: 接入点 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_integration.updateReq'
                description: 更新字段
                summary: body
        description: 更新字段
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Integration'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Not Found
      security:
      - bearerAuth: []
      summary: 更新接入点
      tags:
      - integration
...
  /postmortems:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ent.Postmortem'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 复盘列表
      tags:
      - postmortem
  /postmortems/{id}:
    get:
      parameters:
      - description: 复盘 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Postmortem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Not Found
      security:
      - bearerAuth: []
      summary: 复盘详情
      tags:
      - postmortem
  /postmortems/{id}/action-items:
    post:
      parameters:
      - description: 复盘 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_postmortem.addActionItemReq'
                description: 改进项参数
                summary: body
        description: 改进项参数
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.ActionItem'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 添加改进项
      tags:
      - postmortem
  /postmortems/{id}/transition:
    patch:
      parameters:
      - description: 复盘 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_postmortem.transitionReq'
                description: 目标状态
                summary: body
        description: 目标状态
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Postmortem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
      security:
      - bearerAuth: []
      summary: 复盘状态流转
      tags:
      - postmortem
...
servers:
- url: /api/v1